Added toy files for testing Hooka ttable and vocab generation/conversion process
